ITCollabHub

+++ ITCollabHub/app/src/main/java/ru/dvteam/itcollabhub/CallBackInt4.java
@@ -1,7 +1,7 @@
 package ru.dvteam.itcollabhub;
 
 public interface CallBackInt4 {
-    public void invoke(String name, String photo, String description, int isend, String purposes,
+    public void invoke(String name, String photo, String description, double isend, String purposes,
                        String problems, String peoples, String time, String time1, String tg, String vk, String webs, String purposesids,
                        String problemsids, String isl);
 }
